% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_if_missing.R
\name{clean_if_missing}
\alias{clean_if_missing}
\title{#' Check and clean missing values in a data frame}
\usage{
clean_if_missing(data, method = "remove", fill_value = 0)
}
\arguments{
\item{data}{A data frame.}

\item{method}{How to handle missing values: "remove" (default) or "fill".}

\item{fill_value}{A numeric value used to fill missing numeric values (only used if method = "fill").}
}
\value{
A cleaned data frame.
}
\description{
Automatically checks for missing values. If found, either removes rows or fills values depending on the method.
If no missing values are found, returns the original data with a message.
}
\examples{
clean_if_missing(iris)
clean_if_missing(mtcars, method = "fill", fill_value = -99)
clean_if_missing(mtcars, method = "remove")

}
